archive/performanceanalysis/requirements.txt
apache-beam>=2.61.0
google-cloud-bigtable>=2.26.0
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.13.0
pytest>=7.3.0
pytest-mock>=3.10.0
mypy>=1.0.0
types-all>=1.0.0
black>=22.12.0
isort>=5.12.0
flake8>=6.0.0
